+++ packages/aws-fargate/test/integration_test/app.js
@@ -12,7 +12,20 @@ const port = process.env.TASK_HTTP_PORT || 4215;
 const app = new http.Server();
 
 app.on('request', (req, res) => {
-  fetch(downstreamDummyUrl).then(() => res.end('Hello Fargate!'));
+  fetch(downstreamDummyUrl, {
+    headers: {
+      'X-Exit-Request-Header-1': 'exit request header value 1',
+      'X-Exit-Request-Header-2': ['exit', 'request', 'header', 'value 2'],
+      'X-Exit-Request-Header-3': 'not configured to be captured',
+      'X-Exit-Request-Header-4': ['not', 'configured', 'to', 'be', 'captured']
+    }
+  }).then(() => {
+    res.setHeader('X-Entry-Response-Header-1', 'entry response header value 1');
+    res.setHeader('X-Entry-Response-Header-2', ['entry', 'response', 'header', 'value 2']);
+    res.setHeader('X-Entry-Response-Header-3', 'not configured to be captured');
+    res.setHeader('X-Entry-Response-Header-4', ['not', 'configured', 'to', 'be', 'captured']);
+    res.end('Hello Fargate!');
+  });
 });
 
 app.listen(port, () => {